+++ library/Morloc/CodeGenerator/Serial.hs
@@ -14,6 +14,7 @@ module Morloc.CodeGenerator.Serial
   , isSerializable
   , prettySerialOne
   , serialAstToType
+  , serialAstToType'
   , shallowType
   ) where
 
@@ -24,7 +25,7 @@ import qualified Morloc.Monad as MM
 import qualified Morloc.Data.Text as MT
 import qualified Data.Map as Map
 import qualified Morloc.Frontend.Lang.DefaultTypes as Def
-import Morloc.Pretty (prettyType, prettyPackMap)
+import Morloc.Pretty (prettyType, prettyPackMap, prettyGreenUnresolvedType)
 import Morloc.Data.Doc
 
 typeEqual :: Type -> UnresolvedType -> Bool
@@ -32,6 +33,10 @@ typeEqual (VarT (TV _ v1)) (VarU (TV _ v2)) = v1 == v2
 typeEqual (ArrT (TV _ v1) ts1) (ArrU (TV _ v2) us2)
   | length ts1 /= length us2 = False
   | otherwise = foldl (&&) (v1 == v2) (zipWith typeEqual ts1 us2 )
+typeEqual (NamT (TV _ v1) rs1) (NamU (TV _ v2) rs2)
+  = v1 == v2
+  && map fst rs1 == map fst rs2
+  && foldl (&&) True (zipWith typeEqual (map snd rs1) (map snd rs2))
 typeEqual _ _ = False
 
 -- Convert a default unresolved type to a standard type
@@ -50,13 +55,39 @@ serialAstToType lang (SerialList s) = do
 serialAstToType lang (SerialTuple ss) = do
   ts <- mapM (serialAstToType lang) ss
   return . partialResolve . head . Def.defaultTuple (Just lang) . map partialUnresolve $ ts
-serialAstToType lang (SerialObject v rs) = return $ VarT v
+serialAstToType lang (SerialObject v rs)
+  = return $ NamT v (zip (map fst rs) (map (serialAstToType' lang . snd) rs))
 serialAstToType lang (SerialNum    x) = return $ VarT (TV (Just lang) x)
 serialAstToType lang (SerialBool   x) = return $ VarT (TV (Just lang) x)
 serialAstToType lang (SerialString x) = return $ VarT (TV (Just lang) x)
 serialAstToType lang (SerialNull   x) = return $ VarT (TV (Just lang) x)
-serialAstToType lang (SerialUnknown _) = MM.throwError . SerializationError
-                                       $ "Cannot guess serialization type"
+serialAstToType lang (SerialUnknown x)
+  = MM.throwError . SerializationError . render
+  $ "Cannot guess serialization type:" <+> pretty x
+
+-- | recurse all the way to a serializable type, unsafe
+serialAstToType' :: Lang -> SerialAST One -> Type
+serialAstToType' lang (SerialPack (One (_, s))) = serialAstToType' lang s
+serialAstToType' lang (SerialList s)
+  = partialResolve
+  . head
+  . Def.defaultList (Just lang)
+  . partialUnresolve
+  $ serialAstToType' lang s
+serialAstToType' lang (SerialTuple ss)
+  = partialResolve
+  . head
+  . Def.defaultTuple (Just lang)
+  . map partialUnresolve
+  $ map (serialAstToType' lang) ss
+serialAstToType' lang (SerialObject v rs)
+  = NamT v (zip (map fst rs) (map (serialAstToType' lang . snd) rs))
+serialAstToType' lang (SerialNum    x) = VarT (TV (Just lang) x)
+serialAstToType' lang (SerialBool   x) = VarT (TV (Just lang) x)
+serialAstToType' lang (SerialString x) = VarT (TV (Just lang) x)
+serialAstToType' lang (SerialNull   x) = VarT (TV (Just lang) x)
+serialAstToType' lang (SerialUnknown _) = error "Cannot guess serialization type"
+
 
 -- | get only the toplevel type
 shallowType :: Lang -> SerialAST One -> MorlocMonad Type
@@ -67,7 +98,9 @@ shallowType lang (SerialList s) = do
 shallowType lang (SerialTuple ss) = do
   ts <- mapM (shallowType lang) ss
   return . partialResolve . head . Def.defaultTuple (Just lang) . map partialUnresolve $ ts
-shallowType lang (SerialObject v rs) = return $ VarT v
+shallowType lang (SerialObject v rs) = do
+  ts <- mapM (shallowType lang) (map snd rs)
+  return $ NamT v (zip (map fst rs) ts)
 shallowType lang (SerialNum    x) = return $ VarT (TV (Just lang) x)
 shallowType lang (SerialBool   x) = return $ VarT (TV (Just lang) x)
 shallowType lang (SerialString x) = return $ VarT (TV (Just lang) x)
@@ -107,13 +140,13 @@ makeSerialAST m t@(ArrT v@(TV lang s) ts)
     = SerialList <$> makeSerialAST m (ts !! 0)
   | length tuples > 0 = SerialTuple <$> mapM (makeSerialAST m) ts
   | otherwise = case Map.lookup (v, length ts) m of
-      (Just ps) -> do        
+      (Just ps) -> do
         ps' <- mapM (resolvePacker t ts) ps
         ts' <- mapM (makeSerialAST m) (map typePackerCType ps')
         return $ SerialPack (Many (zip ps' ts'))
       Nothing -> MM.throwError . SerializationError . render
-        $ "Cannot find constructor for" <+> squotes (prettyType t) <+> "in packmap:\n" <>
-          prettyPackMap m 
+        $ "Cannot find constructor" <+> dquotes (pretty s) <> "<" <> pretty (length ts) <> ">" <+> "in packmap:\n" <>
+          prettyPackMap m
   where
     tuples = filter (typeEqual t) (Def.defaultTuple lang (map type2default ts))
 makeSerialAST m (NamT v rs) = do